# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Studio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * JSON canvas state (fabric.js serialization)
 */
export type DesignCanvasData = { [key: string]: unknown };

export interface Design {
  id: number;
  userId: string;
  title: string;
  /** JSON canvas state (fabric.js serialization) */
  canvasData: DesignCanvasData;
  /**
   * Base64 PNG thumbnail
   * @nullable
   */
  thumbnail: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateDesignBodyCanvasData = { [key: string]: unknown };

export interface CreateDesignBody {
  /** @minLength 1 */
  title: string;
  canvasData: CreateDesignBodyCanvasData;
  /** @nullable */
  thumbnail?: string | null;
}

export type UpdateDesignBodyCanvasData = { [key: string]: unknown };

export interface UpdateDesignBody {
  /** @minLength 1 */
  title?: string;
  canvasData?: UpdateDesignBodyCanvasData;
  /** @nullable */
  thumbnail?: string | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};