# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce
    .string()
    .optional()
    .describe(
      "Relative path to redirect to after login (must start with `\/`). Defaults to `\/`.",
    ),
});

/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().url().optional(),
});

/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

/**
 * @summary Exchange a mobile OIDC code for a session token
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().url().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all designs for the current user
 */
export const ListDesignsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  canvasData: zod
    .record(zod.string(), zod.unknown())
    .describe("JSON canvas state (fabric.js serialization)"),
  thumbnail: zod.string().nullable().describe("Base64 PNG thumbnail"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListDesignsResponse = zod.array(ListDesignsResponseItem);

/**
 * @summary Create a new design
 */

export const CreateDesignBody = zod.object({
  title: zod.string().min(1),
  canvasData: zod.record(zod.string(), zod.unknown()),
  thumbnail: zod.string().nullish(),
});

/**
 * @summary Get a design by ID
 */
export const GetDesignParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDesignResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  canvasData: zod
    .record(zod.string(), zod.unknown())
    .describe("JSON canvas state (fabric.js serialization)"),
  thumbnail: zod.string().nullable().describe("Base64 PNG thumbnail"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a design
 */
export const UpdateDesignParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDesignBody = zod.object({
  title: zod.string().min(1).optional(),
  canvasData: zod.record(zod.string(), zod.unknown()).optional(),
  thumbnail: zod.string().nullish(),
});

export const UpdateDesignResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  canvasData: zod
    .record(zod.string(), zod.unknown())
    .describe("JSON canvas state (fabric.js serialization)"),
  thumbnail: zod.string().nullable().describe("Base64 PNG thumbnail"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a design
 */
export const DeleteDesignParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteDesignResponse = zod.object({
  success: zod.boolean(),
});