# lib/api-zod/src/generated/types/createDesignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDesignBodyCanvasData } from "./createDesignBodyCanvasData";

export interface CreateDesignBody {
  /** @minLength 1 */
  title: string;
  canvasData: CreateDesignBodyCanvasData;
  /** @nullable */
  thumbnail?: string | null;
}