# lib/api-zod/src/generated/types/design.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DesignCanvasData } from "./designCanvasData";

export interface Design {
  id: number;
  userId: string;
  title: string;
  /** JSON canvas state (fabric.js serialization) */
  canvasData: DesignCanvasData;
  /**
   * Base64 PNG thumbnail
   * @nullable
   */
  thumbnail: string | null;
  createdAt: Date;
  updatedAt: Date;
}