# lib/api-zod/src/generated/types/updateDesignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QR Studio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDesignBodyCanvasData } from "./updateDesignBodyCanvasData";

export interface UpdateDesignBody {
  /** @minLength 1 */
  title?: string;
  canvasData?: UpdateDesignBodyCanvasData;
  /** @nullable */
  thumbnail?: string | null;
}